feat: v1.0.0 beta

+++ src/main/java/com/airbnl/dbcontrollerservice/model/Reservation.java
@@ -18,8 +18,8 @@ public class Reservation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
-    private int userId;
-    private int roomId;
+    private long userId;
+    private long roomId;
     private double cost;
     private Timestamp dateFrom;
     private Timestamp dateTo;
