+++ src/main/java/com/airbnl/dbcontrollerservice/model/Room.java
@@ -19,6 +19,6 @@ public class Room {
     private long id;
     private long hotelId;
     private String number;
-    private int typeId;
+    private long typeId;
     private int bedCount;
 }
